requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.109.0
h11==0.16.0
idna==3.10
joblib==1.5.1
nltk==3.9.1
numpy==1.26.2
pandas==2.1.4
psycopg2-binary==2.9.9
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.9
pytz==2025.2
regex==2024.11.6
scikit-learn==1.3.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
starlette==0.35.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.27.0
